Move Settings menu items to reflect general purpose
