+++ lib/lightning_web/components/layouts/live.html.heex
@@ -21,29 +21,29 @@
           <LayoutComponents.menu_items {assigns} />
           <div class="grow"></div>
           <%= if @current_user.role == :superuser do %>
-            <Settings.menu_item to={Routes.project_index_path(@socket, :index)}>
+            <Menu.menu_item to={Routes.project_index_path(@socket, :index)}>
               <Heroicons.cog class="h-5 w-5 inline-block mr-2" />
-              <span class="inline-block align-middle">Admin Settings</span>
-            </Settings.menu_item>
+              <span class="inline-block align-middle">Admin Menu</span>
+            </Menu.menu_item>
             <%= if function_exported?(Routes, :live_dashboard_path, 2) do %>
-              <Settings.menu_item to={~p"/dashboard"}>
+              <Menu.menu_item to={~p"/dashboard"}>
                 <Heroicons.exclamation_triangle class="h-5 w-5 inline-block mr-2" />
                 <span class="inline-block align-middle">Sys Info</span>
-              </Settings.menu_item>
+              </Menu.menu_item>
             <% end %>
           <% else %>
-            <Settings.menu_item href="https://docs.openfn.org">
+            <Menu.menu_item href="https://docs.openfn.org">
               <Heroicons.book_open mini class="h-5 w-5 inline-block mr-2" />
               <span class="inline-block align-middle">Documentation</span>
-            </Settings.menu_item>
-            <Settings.menu_item href="https://community.openfn.org">
+            </Menu.menu_item>
+            <Menu.menu_item href="https://community.openfn.org">
               <Heroicons.user_group mini class="h-5 w-5 inline-block mr-2" />
               <span class="inline-block align-middle">Community</span>
-            </Settings.menu_item>
-            <Settings.menu_item href={"mailto:" <> Application.get_env(:lightning, :email_addresses)[:admin]}>
+            </Menu.menu_item>
+            <Menu.menu_item href={"mailto:" <> Application.get_env(:lightning, :email_addresses)[:admin]}>
               <Heroicons.lifebuoy mini class="h-5 w-5 inline-block mr-2" />
               <span class="inline-block align-middle">Support</span>
-            </Settings.menu_item>
+            </Menu.menu_item>
           <% end %>
         <% end %>
         <LightningWeb.Components.Common.version_chip />
