+++ lib/lightning_web/live/components/menu.ex
@@ -1,60 +1,57 @@
 defmodule LightningWeb.Components.Menu do
+  @moduledoc """
+  Menu components to render menu items for project and user/profile pages.
+  """
   use LightningWeb, :component
 
   def project_items(assigns) do
     ~H"""
-    <Settings.menu_item
+    <.menu_item
       to={~p"/projects/#{@project_id}/w"}
       active={@active_menu_item == :overview}
     >
       <Icon.workflows class="h-5 w-5 inline-block mr-2 align-middle" />
       <span class="inline-block align-middle">Workflows</span>
-    </Settings.menu_item>
+    </.menu_item>
 
-    <Settings.menu_item
+    <.menu_item
       to={~p"/projects/#{@project_id}/history"}
       active={@active_menu_item == :runs}
     >
       <Icon.runs class="h-5 w-5 inline-block mr-2" />
       <span class="inline-block align-middle">History</span>
-    </Settings.menu_item>
+    </.menu_item>
 
-    <Settings.menu_item
+    <.menu_item
       to={"/projects/#{@project_id}/settings"}
       active={@active_menu_item == :settings}
     >
       <Icon.settings class="h-5 w-5 inline-block mr-2" />
       <span class="inline-block align-middle">Settings</span>
-    </Settings.menu_item>
+    </.menu_item>
     <!-- # Commented out until new dataclips/globals list is fully functional. -->
     <!--
-      <Settings.menu_item
+      <.menu_item
         to={Routes.project_dataclip_index_path(@socket, :index, @project.id)}
         active={@active_menu_item == :dataclips}>
         <Icon.dataclips class="h-5 w-5 inline-block mr-2" />
         <span class="inline-block align-middle">Dataclips</span>
-      </Settings.menu_item>
+      </.menu_item>
     -->
     """
   end
 
   def profile_items(assigns) do
     ~H"""
-    <Settings.menu_item to={~p"/profile"} active={@active_menu_item == :profile}>
+    <.menu_item to={~p"/profile"} active={@active_menu_item == :profile}>
       <Heroicons.user_circle class="h-5 w-5 inline-block mr-2" /> User Profile
-    </Settings.menu_item>
-    <Settings.menu_item
-      to={~p"/credentials"}
-      active={@active_menu_item == :credentials}
-    >
+    </.menu_item>
+    <.menu_item to={~p"/credentials"} active={@active_menu_item == :credentials}>
       <Heroicons.key class="h-5 w-5 inline-block mr-2" /> Credentials
-    </Settings.menu_item>
-    <Settings.menu_item
-      to={~p"/profile/tokens"}
-      active={@active_menu_item == :tokens}
-    >
+    </.menu_item>
+    <.menu_item to={~p"/profile/tokens"} active={@active_menu_item == :tokens}>
       <Heroicons.command_line class="h-5 w-5 inline-block mr-2" /> API Tokens
-    </Settings.menu_item>
+    </.menu_item>
     """
   end
 
@@ -127,6 +124,47 @@ defmodule LightningWeb.Components.Menu do
           </.link>
         <% end %>
       </ul>
+    </div>
+    """
+  end
+
+  def menu_item(assigns) do
+    base_classes = ~w[px-3 py-2 rounded-md text-sm font-medium rounded-md block]
+
+    active_classes = ~w[text-primary-200 bg-primary-900] ++ base_classes
+
+    inactive_classes = ~w[text-primary-300 hover:bg-primary-900] ++ base_classes
+
+    assigns =
+      assigns
+      |> assign(
+        class:
+          if assigns[:active] do
+            active_classes
+          else
+            inactive_classes
+          end
+      )
+
+    ~H"""
+    <div class="h-12 mx-2">
+      <%= if assigns[:href] do %>
+        <.link href={@href} target="_blank" class={@class}>
+          <%= if assigns[:inner_block] do %>
+            <%= render_slot(@inner_block) %>
+          <% else %>
+            <%= @text %>
+          <% end %>
+        </.link>
+      <% else %>
+        <.link navigate={@to} class={@class}>
+          <%= if assigns[:inner_block] do %>
+            <%= render_slot(@inner_block) %>
+          <% else %>
+            <%= @text %>
+          <% end %>
+        </.link>
+      <% end %>
     </div>
     """
   end
